move command into proper files
