+++ bin/tx-prover/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod init;
+pub mod proxy;
+pub mod worker;
